try setgid approach
OCP allows, I believe, to run containers as non-root users.
The other Dockerfile builds off a branch of stunnel where I simply removed calls to setgid/setuid.
This is an attempt to see if the container and stunnel process run under a non-root users, is the
default behaviour of stunnel to not make calls to setgid/setuid.